Refactor moved generic creaor class to better location
